Created The source and header Files
I added the source files main.cpp and Textprocessor and an header file
TextProcessor.h to the project. A class called TextProcessor was
created also. This will be the base class fioor for the project.
If in case you need to add another class, create a separate
header file for it