Implement COnversion to Uppercase Functionality
In this commit, I have partially implemented conversion to uppercase functionality
using makeAllCaps() function. Right now this function takes the file path to the
text file as input and output all the words converted to uppercase on the terminal.
There can be improvement to this but it will be shared using the issues.